# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: AI CCTV 공개 API — 모바일 교체 계약
  version: 0.3.0
  description: |-
    모바일 등 외부 클라이언트를 교체할 때 유지해야 할 공개 HTTP API이다.
    앱은 중앙 HTTPS origin에 접속하며 Data·Edge·MediaMTX의 내부 포트에 직접 연결하지 않는다.
    paths와 기존 요청·응답 모델은 External의 FastAPI 선언에서 생성한다. 한국어 사용 설명,
    Cookie 인증, Range 헤더 및 Any/dict로 선언된 세 응답은 실제 반환 코드를 근거로 보완한다.
    이 문서는 실행 서버의 /api/v1/openapi.json에 설명을 보강한 배포용 명세다.

    ### 로그인과 권한
    1. POST /api/v1/auth/login으로 access_token·refresh_token과 user를 받는다.
    2. API에는 Authorization: Bearer <access_token> 또는 Access Cookie를 보낸다.
       두 가지가 함께 있으면 Bearer가 우선한다. admin은 전체, viewer는 허용 카메라만 조회한다.
    3. 401이면 현재 refresh_token으로 한 번 갱신하고 원래 요청을 재시도한다.
       갱신은 새 access·refresh 쌍을 반환하며 이전 refresh는 재사용할 수 없다.
       앱은 동시 갱신을 하나로 합치고 최신 두 토큰을 함께 저장한다. 갱신 실패는 재로그인한다.
    4. 로그아웃에는 현재 access와 refresh를 함께 보낸다. 서버는 제출된 두 토큰을 폐기하고
       refresh 계열의 기기 등록·대기 알림과 쿠키를 삭제한다. 다른 access까지 전부 즉시 철회하는
       전체 기기 로그아웃 API는 아니다. 이미 회전된 refresh 대신 반드시 최신값을 보낸다.
    기본 쿠키명은 ai_cctv_access, ai_cctv_refresh이며 배포 설정으로 바뀔 수 있다.
    로그인·갱신은 HttpOnly 쿠키도 설정한다. 기본 Secure=true, SameSite=lax,
    access Path=/, refresh Path=/api/v1/auth, 기본 수명은 각각 900초·604800초이다.
    실제 access 수명은 expires_in을 따른다. 토큰을 URL·로그에 기록하지 않는다.

    ### 영상 재생: Nginx 경로
    GET /cameras/{camera_id}/live의 url 또는 GET /recordings/{id}/playback의 playback_url을
    그대로 사용한다. PUBLIC_BASE_URL 설정 시 HTTPS 절대 URL, 없으면 API origin 기준 상대 URL이다.
    라이브 기본 주소는 /hls/{camera_id}/index.m3u8이고 재생목록과 영상 조각을 받는 HLS이다.
...

...

...

...
paths:
...
  /api/v1/auth/login:
    post:
      summary: 로그인
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          headers:
            Set-Cookie:
              description: access·refresh 쿠키를 각각 설정한다. 수명·경로·기본 이름은 위 인증 설명 참조.
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: 반복 로그인 실패로 일시 지연됨
          headers:
            Retry-After:
              description: 다시 시도하기 전 기다릴 초
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailError'
      description: 계정 확인 후 토큰 쌍과 쿠키를 발급한다. 반복 실패 시 429와 Retry-After를 반환한다.
      tags:
      - auth
      security: []
  /api/v1/auth/refresh:
    post:
      summary: 토큰 갱신
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/RefreshRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          headers:
            Set-Cookie:
              description: access·refresh 쿠키를 각각 설정한다. 수명·경로·기본 이름은 위 인증 설명 참조.
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      description: Body의 refresh_token이 쿠키보다 우선한다. 둘 중 하나가 필요하다. 성공하면 이전 refresh를 폐기하므로 동시 갱신을 피한다.
      tags:
      - auth
      security: []
      parameters:
      - in: cookie
        name: ai_cctv_refresh
        required: false
        schema:
          type: string
        description: Body의 refresh_token이 없을 때 사용한다. REFRESH_COOKIE_NAME 설정으로 이름이 바뀔 수 있다.
  /api/v1/auth/logout:
    post:
      summary: 로그아웃
      operationId: logout_api_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/LogoutRequest'
              - type: 'null'
              title: Payload
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ApiError'
      description: Body 또는 쿠키의 현재 refresh와 Authorization 또는 쿠키의 access를 함께 보낸다. 유효한 토큰만 철회하며 토큰이 없거나 만료된
        경우에도 쿠키 삭제 후 204일 수 있다. refresh 철회는 그 계열의 기기·대기 알림도 해제한다. 이미 FCM에 접수된 알림은 취소되지 않는다.
      tags:
      - auth
      security:
      - HTTPBearer: []
      - AccessCookie: []
      - {}
      parameters:
      - in: cookie
        name: ai_cctv_refresh
        required: false
        schema:
          type: string
        description: Body의 refresh_token이 없을 때 사용한다. REFRESH_COOKIE_NAME 설정으로 이름이 바뀔 수 있다.
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 128
          minLength: 1
          pattern: ^[^/\\\x00-\x1f]+$
          title: Username
        password:
          type: string
          maxLength: 1024
          minLength: 1
          format: password
          title: Password
          writeOnly: true
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: LoginRequest
    LogoutRequest:
      properties:
        refresh_token:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Refresh Token
      additionalProperties: false
      type: object
      title: LogoutRequest
...
    RefreshRequest:
      properties:
        refresh_token:
          anyOf:
          - type: string
            format: password
            writeOnly: true
          - type: 'null'
          title: Refresh Token
      additionalProperties: false
      type: object
      title: RefreshRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          const: bearer
          title: Token Type
        expires_in:
          type: integer
          exclusiveMinimum: 0.0
          title: Expires In
        user:
          $ref: '#/components/schemas/UserResponse'
      type: object
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      - user
      title: TokenResponse
...
    UserResponse:
      properties:
        id:
          anyOf:
          - type: integer
          - type: string
          title: Id
        username:
          type: string
          title: Username
        role:
          type: string
          enum:
          - admin
          - viewer
          title: Role
        is_active:
          type: boolean
          title: Is Active
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - username
      - role
      - is_active
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    DetailError:
      type: object
      properties:
        detail:
          type: string
      required:
      - detail
    ControlError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
          - code
          - message
          - details
      required:
      - error
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access_token. refresh_token을 Bearer로 사용하지 않는다.
    AccessCookie:
      type: apiKey
      in: cookie
      name: ai_cctv_access
      description: 기본 access 쿠키명. 배포의 ACCESS_COOKIE_NAME으로 변경할 수 있다.
  responses:
    ApiError:
      description: 요청·권한·상태·상위 서비스 오류. 일반 detail 또는 제어 error 구조이며 422는 detail 배열이다. 프록시나 처리하지 못한 서버 오류는
        JSON이 아닐 수 있다.
      content:
        application/json:
          schema:
            anyOf:
            - $ref: '#/components/schemas/DetailError'
            - $ref: '#/components/schemas/ControlError'
            - $ref: '#/components/schemas/HTTPValidationError'
        text/plain:
          schema:
            type: string
        text/html:
          schema:
            type: string
    Unauthorized:
      description: 인증 정보가 없거나 만료·철회·변경되었다. 현재 refresh로 한 번만 갱신하고 실패하면 다시 로그인한다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DetailError'
servers:
...